Add scene detail header
